+++ package.json
@@ -7,12 +7,14 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.4.0",
+    "chai": "^4.3.7",
     "cheerio": "^1.0.0-rc.12",
     "cors": "^2.8.5",
     "cors-anywhere": "^0.4.4",
     "express": "^4.18.2",
     "install": "^0.13.0",
     "jsdom": "^22.1.0",
+    "mocha": "^10.2.0",
     "npm": "^9.7.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
