+++ tsconfig.json
@@ -15,5 +15,5 @@
     "noEmit": true,
     "jsx": "react",
   },
-  "include": ["src"]
+  "include": ["src", "tests"]
 }
